Cleaning up library refactor
