+++ lib/elemeno.js
@@ -1,105 +1,118 @@
-const request = require('superagent'),
-	baseUrl = 'https://api.elemeno.io/v1/',
-	singleBase = 'singles/',
-	collectionBase = 'collections/';
+"use strict";
 
+const request = require('superagent');
 
-function Elemeno(apiKey) {
+class Elemeno {
 
-}
+	constructor (apiKey) {
+		this.apiKey = apiKey;
+		this.baseUrl = 'https://api.elemeno.io/v1/';
+		this.singleBase = 'singles/';
+		this.collectionBase = 'collections/';
+  	}
 
-var apiKey;
+	static getQuery(options, allow) {
+		let query = {};
 
-function get(path, query, cb) {
-	request
-	.get(baseUrl + path)
-	.set('Authorization', apiKey)
-	.query(query)
-	.end(function(err, res) {
-		return (err) ? cb(err.response.error) : cb(null, res.body);
-	});
-}
-
-function getQuery(options, allow) {
-	var query = {};
-
-	if (options) {
-		if (options.filters && Array.isArray(allow) && allow.indexOf('filters') > -1) {
-			query.filters = JSON.stringify(options.filters);
+		if (options) {
+			if (options.filters && Array.isArray(allow) && allow.indexOf('filters') > -1) {
+				query.filters = JSON.stringify(options.filters);
+			}
+			if (options.sort && Array.isArray(allow) && allow.indexOf('sort') > -1) {
+				query.sort = JSON.stringify(options.sort);
+			}
+			if (options.page) {
+				query.page = options.page;
+			}
+			if (options.size) {
+				query.size = options.size;
+			}
+			if (options.byId) {
+				query.byId = (options.byId === true || options.byId === 'true') ? true : false;
+			}
 		}
-		if (options.sort && Array.isArray(allow) && allow.indexOf('sort') > -1) {
-			query.sort = JSON.stringify(options.sort);
-		}
-		if (options.page) {
-			query.page = options.page;
-		}
-		if (options.size) {
-			query.size = options.size;
-		}
-		if (options.byId) {
-			query.byId = (options.byId === true || options.byId === 'true') ? true : false;
-		}
+
+		return query;
 	}
 
-	return query;
-}
+	get(path, query, cb) {
+		if (typeof cb === 'undefined') {
+			cb = query;
+			query = null;
+		}
 
-var elemeno = {};
+		let req = request.get(this.baseUrl + path);
 
-elemeno.setAPIKey = function(key) {
-	apiKey = key;
-}
+		req.set('Authorization', this.apiKey);
 
-// Singles
+		if (query) {
+			req.query(query)
+		}
 
-elemeno.getSingles = function(options, cb) {
-	if (typeof cb === 'undefined') {
-		cb = options;
-		options = {};
+		req.end(function(err, res) {
+			return (err) ? cb(err.response.error) : cb(null, res.body);
+		});
 	}
 
-	return get(singleBase, getQuery(options, ['sort']), cb);
-}
+	getSingles(options, cb) {
+		if (typeof cb === 'undefined') {
+			cb = options;
+			options = null;
+		}
 
-elemeno.getSingle = function(singleSlug, cb) {
-	var options = {};
+		let path = this.singleBase;
+		let query = Elemeno.getQuery(options, ['sort']);
 
-	return get(singleBase + singleSlug, getQuery(options), cb);
-}
-
-// Collections
-
-elemeno.getCollections = function(options, cb) {
-	if (typeof cb === 'undefined') {
-		cb = options;
-		options = {};
+		return this.get(path, query, cb);
 	}
 
-	return get(collectionBase, getQuery(options, ['sort']), cb);
-}
+	getSingle(singleSlug, cb) {
+		let path = this.singleBase + singleSlug;
 
-elemeno.getCollection = function(collectionSlug, cb) {
-	var options = {};
-
-	return get(collectionBase + collectionSlug, getQuery(options), cb);
-}
-
-elemeno.getCollectionItems = function(collectionSlug, options, cb) {
-	if (typeof cb === 'undefined') {
-		cb = options;
-		options = {};
+		return this.get(path, cb);
 	}
 
-	return get(collectionBase + collectionSlug + '/items/', getQuery(options, ['filters', 'sort']), cb);
-}
+	getCollections(options, cb) {
+		if (typeof cb === 'undefined') {
+			cb = options;
+			options = null;
+		}
 
-elemeno.getCollectionItem = function(collectionSlug, itemSlug, options, cb) {
-	if (typeof cb === 'undefined') {
-		cb = options;
-		options = {};
+		let path = this.collectionBase;
+		let query = Elemeno.getQuery(options, ['sort']);
+
+		return this.get(path, query, cb);
 	}
 
-	return get(collectionBase + collectionSlug + '/items/' + itemSlug, getQuery(options), cb);
+	getCollection(collectionSlug, cb) {
+		let path = this.collectionBase + collectionSlug;
+
+		return this.get(path, cb);
+	}
+
+	getCollectionItems(collectionSlug, options, cb) {
+		if (typeof cb === 'undefined') {
+			cb = options;
+			options = null;
+		}
+
+		let path = this.collectionBase + collectionSlug + '/items/';
+		let query = Elemeno.getQuery(options, ['filters', 'sort']);
+
+		return this.get(path, query, cb);
+	}
+
+	getCollectionItem(collectionSlug, itemSlug, options, cb) {
+		if (typeof cb === 'undefined') {
+			cb = options;
+			options = null;
+		}
+
+		let path = this.collectionBase + collectionSlug + '/items/' + itemSlug;
+		let query = Elemeno.getQuery(options);
+
+		return this.get(path, query, cb);
+	}
 }
 
-module.exports = elemeno;
+module.exports = Elemeno;
